ESS-26: swap vim-tiny for nano
Users who want vim will have the nous to install vim themselves, for everyone
else the default editor should be nano. Once you install standard with nano
as a dependency, the default editor would be nano anyway (and vim-tiny would
just be left around), so we just move it back a step.