ESS-35: remove kerneloops-daemon since it was depending on apport
kerneloops-daemon talks to kerneloops.org which might be useful for
mainline kernels but they're not going to get a lot of great info
from us or appreciate things going there.

We want to get rid of apport because it uploads the same kind of bug
reports to Launchpad which is not friendly to the Ubuntu guys, so
this can go.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>