ESS-25: create genesi-imx5-extras package
We can't use Germinate because it doesn't properly handle packages
with a dash in their name (perl split function) and therefore creates
no substitution variables since it expects that the architecture is the
first thing after the dash.

Maybe we should file a bug upstream, since the arch should be the last
thing after the LAST dash..

Ditch a couple redundant files while we are here.

Bump to 1.214 to cover ESS-25.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>