ESS-69/ESS-52: create genesi-office package
We should pull this by hand, in case we actually want to remove it
from builds or add extra office-type packages into the mix (like
openoffice plugins or specific language support).

This removes the dependency on openoffice.org from the desktop
system, while not actually having a hard dependency on a specific
office variant for a specific desktop (as standard ubuntu has
openoffice.org as a dependency, xubuntu has abiword/gnumeric,
which means mixing and matching causes a mess. We don't want
that mess here..)

Signed-off-by: Matt Sealey <matt@genesi-usa.com>